Promote SHA-256 + HMAC-SHA256 to a shared crypto module + add RFC KATs
Next TLS Phase 0 sub-step (per docs/audit/TLS-RECON.md §1): lift the file-static
SHA-256 (FIPS 180-4) and HMAC-SHA256 (RFC 2104) out of middleware_auth.c into a
shared src/crypto/sha256.{c,h} so the JWT verifier and the upcoming pure-C TLS
key schedule share one implementation instead of duplicating it.

- The implementation is moved VERBATIM (byte-for-byte identical logic); the only
  change is linkage: the public entry points (sha256_init/update/final, sha256,
  hmac_sha256) are no longer static, while the round macros / K constants /
  transform remain file-local in sha256.c.
- src/crypto/sha256.c joins WEBLIB_SOURCES_WASM_SAFE (pure logic, no OS I/O) — it
  is compiled into every target, since middleware_auth.c uses it on all of them.
  Not part of the public API (kamran.k); internal to src/.
- middleware_auth.c now #includes "crypto/sha256.h" and drops its copy.

Tests: added test_sha256_kat (RFC 6234 vectors: "abc", empty, the 56-byte string,
plus a streaming-equals-one-shot check) and test_hmac_sha256_kat (RFC 4231 TC1/TC2
and TC6's 131-byte hash-the-key-first case). Until now this crypto was only
exercised INDIRECTLY through JWT signature checks; it is now verified directly
against the authoritative vectors. Negative control (corrupt one SHA-256 init
constant) fails the SHA KAT, the HMAC KAT, AND JWT verification — confirming the
KATs are load-bearing and that JWT routes through the shared module.

Validation: ctest 6/6 on normal and ASan/UBSan (default build), 7/7 on the
TLS-enabled build; zero warnings; JWT behavior unchanged.
